Update skeleton.py
Accounting for multiple intersections between downstream LineString and clean skeleton in final_skeleton.

Removing the possibility to use a Polygon for the channel network. It must be a MultiPolygon (of only 1 Polygon if needed). Accordingly, in the entire module, skeletons are always list of skeletons are lists of MultiLineStrings and downstream LineStrings are MultiLineStrings.